Combined image scaling to ocr script
