+++ ocr.py
@@ -1,13 +1,25 @@
 #!/usr/bin/env python
 import requests
 import time
+import io
+from PIL import Image
+
+def scaleImage(imgFile, maxSize):
+    size = maxSize, maxSize
+    img = Image.open(imgFile)
+    imgWithinLimit = all(dimension < limit for dimension, limit in zip(img.size, size))
+    if (not imgWithinLimit):
+        img.thumbnail(size, Image.ANTIALIAS)
+    imgBytes = io.BytesIO()
+    img.save(imgBytes, format="JPEG")
+    return imgBytes
 
 subscription_key = "e4f9a2574d75482597cd04bb93cfb3e0"
 vision_base_url = "https://westcentralus.api.cognitive.microsoft.com/vision/v1.0/"
 text_recognition_url = vision_base_url + "recognizeText"
 
-image_path = "testImages/scrumboard.jpg"
-image_data = open(image_path, 'rb').read()
+image_path = "testImages/bigScrumboard.jpg"
+image_data = scaleImage(image_path, 3200).getvalue()
 
 headers    = {
     'Ocp-Apim-Subscription-Key': subscription_key,
